+++ src/proc_mapping/region_of_interest/rotated_rectangle.h
@@ -53,7 +53,7 @@ class RotatedRectangle : public RegionOfInterest {
 
   explicit RotatedRectangle(const YAML::Node &node);
   explicit RotatedRectangle(const std::string &name, const cv::Point2d &center,
-                            const cv::Size2f &size, double angle,
+                            const cv::Size2f &size, int angle,
                             const DetectionMode &mode = DetectionMode::NONE);
 
   virtual ~RotatedRectangle() = default;
@@ -84,7 +84,7 @@ class RotatedRectangle : public RegionOfInterest {
 
   cv::Size2f size_;
 
-  double angle_;
+  int angle_;
 };
 
 }  // namespace proc_mapping
